requirements.txt
telethon
python-dotenv
cryptg
httpx
pytz
fastapi
uvicorn
jinja2
emoji
